requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
numpy==1.26.4
scikit-learn==1.4.2
pydantic==2.7.1
requests==2.31.0
rank-bm25==0.2.2
loguru==0.7.2
pandas==2.2.2
matplotlib==3.8.4
umap-learn==0.5.6